refactor(model-list): clarify runtime key fallback naming
