+++ src/features/ModelList/hooks/useModelListData.ts
@@ -275,7 +275,8 @@ export function useModelListData(routeParams?: Record<string, string>) {
     sourceCapabilities,
     isFallbackCatalogActive,
     isAihubmixCatalogFallbackActive,
-    fallbackTokenName: modelData.accountFallback?.activeRuntimeKeyName ?? null,
+    fallbackRuntimeKeyName:
+      modelData.accountFallback?.activeRuntimeKeyName ?? null,
 
     ...state,
     ...modelData,
